+++ src/base/enums.c
@@ -36,10 +36,17 @@ static const key_type_value_t align_v_name_value[] = {
     {"top", 0, ALIGN_V_TOP}, {"middle", 0, ALIGN_V_MIDDLE}, {"bottom", 0, ALIGN_V_BOTTOM}};
 
 static const key_type_value_t input_type_name_value[] = {
-    {"int", 0, INPUT_INT},       {"float", 0, INPUT_FLOAT}, {"uint", 0, INPUT_UINT},
-    {"ufloat", 0, INPUT_UFLOAT}, {"text", 0, INPUT_TEXT},   {"password", 0, INPUT_PASSWORD},
-    {"hex", 0, INPUT_HEX},       {"email", 0, INPUT_EMAIL}, {"phone", 0, INPUT_PHONE},
+    {"int", 0, INPUT_INT},
+    {"float", 0, INPUT_FLOAT},
+    {"uint", 0, INPUT_UINT},
+    {"ufloat", 0, INPUT_UFLOAT},
+    {"text", 0, INPUT_TEXT},
+    {"password", 0, INPUT_PASSWORD},
+    {"hex", 0, INPUT_HEX},
+    {"email", 0, INPUT_EMAIL},
+    {"phone", 0, INPUT_PHONE},
     {"custom", 0, INPUT_CUSTOM},
+    {"custom_password", 0, INPUT_CUSTOM_PASSWORD},
 };
 
 static const key_type_value_t align_h_name_value[] = {